fix a bug of rv32i.py

+++ rafi/rv32i.py
@@ -525,7 +525,7 @@ class ADD(Op):
 
     def execute(self, cpuState, bus):
         x = cpuState.int_reg
-        x[self.rd] = x[self.rs1] + x[self.rs2]
+        x[self.rd] = UInt32(x[self.rs1] + x[self.rs2])
 
 class SUB(Op):
     def __init__(self, rd, rs1, rs2):
